+++ src/pages/Delete/Delete.scss
