+++ src/pages/Inbody/Inbody.scss
