+++ src/pages/insertHealth/insertHealth.scss
@@ -89,7 +89,7 @@
                 .fat-data{
                     font-size: large;
                     width:200px;
-                    margin: 0px 0px 0px 30px;
+                    margin: 0px 0px 0px 27px;
                 }
             }
